refactor: separate task state from snapshot publishing
